+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
-## Unreleased
+## v3.0.0
 
 ### Added
 
 - File cleaning configuration has been moved to `.clean_rules.php`
+- Support for rules that exclude deletion of files and packages
 
 ### Changed